denndaku

+++ src/main/java/com/example/demo/week5/service/CompoundInterestService.java
@@ -20,12 +20,9 @@ public class CompoundInterestService {
      */
 	
 	public double getResult(int principal,double rate,int years) {
-
 		double Rate = Math.pow(1 + rate / 100, years);
-		double result = (double)(principal * Rate);
-			
-			return result;
-		
+		double result = (principal * Rate);
+		return result;
 	}
 }
